+++ addon/helpers/-base.js
@@ -16,7 +16,14 @@ export default Helper.extend({
     this.clearTimer();
 
     if (interval) {
-      this.intervalTimer = run.later(this, 'recompute', parseInt(interval, 10));
+      /*
+       * NOTE: intentionally a setTimeout so tests do not block on it
+       * as the run loop queue is never clear so tests will stay locked waiting
+       * for queue to clear.
+       */
+      this.intervalTimer = setTimeout(() => {
+        run(() => this.recompute());
+      }, parseInt(interval, 10));
     }
   },
 
@@ -38,7 +45,7 @@ export default Helper.extend({
   },
 
   clearTimer() {
-    run.cancel(this.intervalTimer);
+    clearTimeout(this.intervalTimer);
   },
 
   destroy() {