+++ blueprints/ember-moment/index.js
@@ -8,6 +8,6 @@ module.exports = {
   },
 
   afterInstall: function() {
-    return this.addAddonToProject('ember-cli-moment-shim', '^1.2.0');
+    return this.addAddonToProject('ember-cli-moment-shim', '^3.0.0');
   }
 };