+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 7.0.3
+
+* Upgrade ember-cli-moment-shim to 3.0.0
+* helper recompute loop now inside of setTimeout so tests do not hang waiting on queue to clear #205
+
 ### 7.0.2
 
 * Revert upgrade ember-cli-moment-shim to 2.2.1 (moment.now issue)
@@ -14,7 +19,7 @@
 * Added `moment-calendar` helper
 * @GarPit Added `moment-to-date` helper
 * @stavarotti added `unix` helper
-* @yads added `moment-subtract` `moment-add` helpers 
+* @yads added `moment-subtract` `moment-add` helpers
 * Removed `ember-getowner-polyfill` as a direct dep
 * Add moment query helpers (is-between, is-sae-or-after, is-same-or-before, is-same, is-after, is-before)
 * BREAKING: no longer defaults moment-format to LLLL and instead uses the moment default format (ISO 8601)